Initial pass on adding support for persistence using sqlcipher
To keep the database schema simple we just save the
otpauth:// url and let KeyInfo take care of checking for
validity.

This patch also includes multiple fixes for the KeyInfo
class.

We still need a separate activity to allow the user to
enter their PIN/password. Currently, "test" is used as the
password for the database.