Replace SpongyCastle with BouncyCastle
SpongyCastle is a fork of BouncyCastle. We originally used this fork to 1) have
access to scrypt and 2) prevent a package name collision with the bundled
BouncyCastle. We don't actually need to use the fork anymore, because the
package name of the bundled BouncyCastle was changed in Android. SpongyCastle
has also gotten quite outdated in recent years.

This also replaces the built-in version of BouncyCastle with the one bundled
with the app, so that we have a recent version even on older Android versions.
Recent versions of BouncyCastle also provide Argon2, which a KDF I would like to
experiment with as a replacement for scrypt.