Switch to libsodium for the scrypt implementation
This drops the dependency we have on SpongyCastle for the implementation of
scrypt in favor of the one in libsodium. With this, we fix crashes on low-end
devices (#114, #300), get better scrypt performance and pave the way for
switching to better crypto primitives in the future as discussed in
[vault.md](https://github.com/beemdevelopment/Aegis/blob/master/docs/vault.md#primitives).

Libsodium is a C library. I wrote a small JNI binding library for it
([libsodium-android](https://github.com/beemdevelopment/libsodium-android)), as
I wasn't happy with the existing options. I placed the .AAR file in source
control for now, as I'm not sure where to publish it. I considered GitHub's
maven package repository, but apparently it always requires authentication, even
for downloading packages. Let's test this first and figure out where we should
publish libsodium-android later.