Update dependencies and enable dependency verification in Gradle
Here's another attempt to enable Gradle dependency verification. I've
tried to do this a couple of times before, but always ran into issues.
With a helper script borrowed from a different project, hopefully this
will finally work reliably.

Every time we update Gradle or a dependency, the script I mentioned
needs to be run to update the list of checksums. The script runs the
build, the tests and the instrumented tests. This is necessary
because apparently Gradle is not able to discover all dependencies
without actually running the build tasks.

It's a little bit annoying that this Gradle feature is so half-baked
that we need a helper script to make things work, but I think it's worth
it.